Remove warnings (trailing \, signed vs unsigned, parenthesization). Convert some tabs into spaces.
git-svn-id: http://skia.googlecode.com/svn/trunk@2541 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/BlurTest.cpp
@@ -121,7 +121,7 @@ static void test_blur(skiatest::Reporter* reporter) {
             paint.setMaskFilter(filter);
             filter->unref();
 
-            for (int test = 0; test < SK_ARRAY_COUNT(tests); ++test) {
+            for (size_t test = 0; test < SK_ARRAY_COUNT(tests); ++test) {
                 SkPath path;
                 tests[test].addPath(&path);
                 SkPath strokedPath;
